add users list to admin panel + some roles logic updates

+++ app/Http/Controllers/Backend/AdminVendorController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers\Backend;
 
 use App\DataTables\VendorDataTable;
 use App\Http\Controllers\Controller;
+use App\Models\Product;
+use App\Models\User;
 use App\Models\Vendor;
+use App\Models\Wishlist;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
 
@@ -36,8 +39,9 @@ class AdminVendorController extends Controller
         $request->validate([
             'banner' => ['required', 'image', 'max:2048'],
             'name' => ['required','string', 'max:200'],
+            'user_id' => ['required', 'integer', 'exists:users,id'],
             'phone' => ['required', 'numeric'],
-            'email' => ['Email', 'max:200'],
+            'email' => ['Email', 'max:200', 'unique:vendors,email'],
             'address' => ['required','string', 'max:200'],
             'description' => ['required','string', 'max:1000'],
             'fb_link' => ['url'],
@@ -45,7 +49,9 @@ class AdminVendorController extends Controller
             'insta_link' => ['url'],
             'status' => ['required', 'boolean'],
         ]);
-
+        $user = User::findOrFail($request->user_id);
+        $user->role = 'vendor';
+        $user->save();
         $vendor = new Vendor();
         $imagePath = $this->uploadImage($request, 'banner', 'uploads');
         $alert = 'New Vendor Has Been Created!';
@@ -82,6 +88,7 @@ class AdminVendorController extends Controller
         $request->validate([
             'banner' => ['nullable', 'image', 'max:2048'],
             'name' => ['required','string', 'max:200'],
+            'user_id' => ['required', 'integer', 'exists:users,id'],
             'phone' => ['required', 'numeric'],
             'email' => ['Email', 'max:200'],
             'address' => ['required','string', 'max:200'],
@@ -92,7 +99,31 @@ class AdminVendorController extends Controller
             'status' => ['required', 'boolean'],
         ]);
 
+        $user = User::findOrFail($request->user_id);
         $vendor = Vendor::findOrFail($id);
+        if($vendor->user_id != $user->id){
+            $vendor->status = 0;
+            $vendor->is_approved = 0;
+            $vendor->save();
+            //check if vendor has another vendor instances
+            $vendors = Vendor::where('user_id', $vendor->user->id)->get();
+            if($vendors->isEmpty()){
+                $user->role = 'user';
+                $user->save();
+                return response(['status' => 'success', 'message' => 'Vendor Is Deleted!']);
+            }
+            $products = Product::where('vendor_id', $vendor->id)->get();
+            if($products->isNotEmpty()){
+                foreach ($products as $product){
+                    $product->status = 0;
+                    $product->is_approved = 0;
+                    $product->save();
+                }
+            }else{
+                $vendor->delete();
+            }
+        }
+
         $imagePath = $this->updateImage($request, 'banner', 'uploads', $vendor->banner);
         $alert = 'Vendor Has Been Updated!';
         $route = 'admin.vendor.index';
@@ -106,14 +137,37 @@ class AdminVendorController extends Controller
     public function destroy(string $id)
     {
         $vendor = Vendor::findOrFail($id);
-        $vendor->delete();
-
-        return response(['status' => 'success', 'message' => 'Vendor Is Deleted!']);
+        $products = Product::where('vendor_id', $vendor->id)->get();
+        if($products->isNotEmpty()){
+            return response(['status' => 'error', 'message' => 'This Vendor Has Registered Products! Delete All Vendor Products to Delete This User.']);
+        }else{
+            $user = User::findOrFail($vendor->user->id);
+            $vendors = Vendor::where('user_id', $user->id)->get();
+            if($vendors->isEmpty()){
+                $user->role = 'user';
+                $user->save();
+            }
+            $vendor->delete();
+            return response(['status' => 'success', 'message' => 'Vendor Is Deleted!']);
+        }
     }
 
     public function changeStatus(Request $request){
         $vendor = Vendor::findOrFail($request->id);
-        $vendor->status = $request->status == 'true' ? 1 : 0;
+        $products = Product::where('vendor_id', $vendor->id)->get();
+        if($request->status === 'true'){
+            $vendor->status = 1;
+            foreach ($products as $product){
+                $product->status = 1;
+                $product->save();
+            }
+        }else if($request->status === 'false'){
+            $vendor->status = 0;
+            foreach ($products as $product){
+                $product->status = 0;
+                $product->save();
+            }
+        }
         $vendor->save();
 
         return response(['message' => 'Status Has Been Changed']);
@@ -132,6 +186,7 @@ class AdminVendorController extends Controller
         $vendor->tw_link = $request->tw_link;
         $vendor->insta_link = $request->insta_link;
         $vendor->status = $request->status;
+        $vendor->is_approved = 1;
         $vendor->save();
 
         toastr()->success($alert);
